requirements.txt
python-docx==24.0
pyttsx==2.90